# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GymFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all membership plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"

    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a plan by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found

    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update a plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"

    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Subscriptions ─────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - in: query
          name: memberId
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Notifications ─────────────────────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"

    post:
      operationId: createStaff
      tags: [staff]
      summary: Add a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Remove a staff member
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        gymId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationDays:
          type: integer
        benefits:
          type: ["string", "null"]
        isActive:
          type: boolean
        memberCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - gymId
        - name
        - price
        - durationDays
        - isActive
        - memberCount
        - createdAt
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationDays:
          type: integer
        benefits:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
        - price
        - durationDays
    UpdatePlanBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationDays:
          type: integer
        benefits:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Subscriptions ─────────────────────────────────────────────────────
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: ["string", "null"]
        subscriptionId:
          type: ["integer", "null"]
        amount:
          type: number
        method:
          type: string
          enum: [cash, card, upi, bank_transfer, other]
        notes:
          type: ["string", "null"]
        paidAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - memberId
        - amount
        - method
        - paidAt
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        memberId:
          type: integer
        subscriptionId:
          type: ["integer", "null"]
        amount:
          type: number
        method:
          type: string
          enum: [cash, card, upi, bank_transfer, other]
        notes:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
      required:
        - memberId
        - amount
        - method

    # ── Notifications ─────────────────────────────────────────────────────
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        gymId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, staff]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - gymId
        - name
        - email
        - role
        - isActive
        - createdAt
    CreateStaffBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, staff]
      required:
        - name
        - email
        - password
        - role
    UpdateStaffBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, staff]
        isActive:
          type: boolean